build: Move all ramsey protocols in ramsey folder
